initialize subscriptions lazily and reconnect upon max poll exceeded
Consider a consumer (or producer) with two subscriptions and one of them being very
full. Upon boot, we'd subscribe to all topics but then only work on the full one,
since we only ever switch when a poll doesn't return a message. However, all topics
are subject to `max.poll.interval.ms` (5 minutes by default) in newer librdkafka
versions and so racecar will crash whenever it starves a topic for more than this
timeout.

The patch solves this through lazy initialization and reconnect when this error
occurrs. Ideally racecar would disconnect starved topics on its own.