manually set the offset in favor of manually committing
The advantage is that setting the offset is fast, since only a single
value is updated. It is committed asynchronously in the background.
This was suggested both from rdkafka-ruby [1] and librdkafka [2].

This might also fix the issue where offsets "disappear". It's suspected
this has to do with not receiving new messages for longer than
offsets.retention.minutes and thus not committing. Even though the
consumer stays connected, the offsets get cleaned up. With the patch
there will be auto commits on an interval, which might stop the
offsets from being cleaned up.

[1] https://github.com/appsignal/rdkafka-ruby/issues/57#issuecomment-474737751
[2] within max.poll.interval.ms: https://github.com/edenhill/librdkafka/blob/d0195a0fdd35aca12e4d4277ca3ae11d6165da27/CONFIGURATION.md