manually group batched messages per partition
librdkafka passes whichever messages are available to poll, even
if there are still messages available for the current partition.
This means that listening for the EOF partition is not helpful,
because one can end up with multiple partitions per batch anyway.

To work around this, it simply groups the received batch per
partition and calls the consumer individually for each group.

Right now the batch limit is enforced for racecar, i.e. no
more than n messages are in our memory at a time. Arguably
the batch size limit could also be per consumer and there'd
be an upper limit of (n-1)*partitions messages kept, while
still keeping to the consumer's desired batch size. In an eyeball
test it appears the "multi partition batches" appear in about 10%
of the cases, so this is a potential performance optimization for
later.