fix race condition when rescue block is slow
The issue was that something in the rescue block could exceed the remaining
time budget, which would be evaluated in the main block. If it was 0 there,
we would just return nil when the method should've raised.

Now, the remaining time budget is calculated at the end of the rescue block,
which solves the "not raising" race condition.