Track number of retries when handling errors
This commit introduces a new `retries_count` key in the `info` hash
passed to the configured error handler.

This can be useful in cases where we might want to limit the number of
individual error occurrences that get reported to a exception
notification service or even implement some sort of monitoring for
errors.