Add statistics_interval to config
NOTE: this can be pulled out into its own PR! Just testing alongside the
forking changes for now.

Statistics are currently provided by default to applications using
Racecar, however, this is not much use if no callback is defined (i.e.
if nobody is listening). It's currently a source of memory bloat, as a
large-ish hash (at least for the consumer) is generated each second.