Restore support for Ruby 2.4
The "ensure inside a block without an explicit begin/end" feature
is a Ruby 2.5 feature (see <https://bugs.ruby-lang.org/issues/12906>)
and its usage broke racecar usage on Ruby 2.4 (#252).

By re-adding the begin/end, it all works again. To make sure it
keeps working, I've also added Ruby 2.4 to the test matrix on
GitHub.

I also added a `spec.required_ruby_version` to the gemspec, to make
it clear which versions are tested. I tried checking if Ruby 2.3 or
older could be trivially supported, but in fact the `rdkafka`
dependency also requires 2.4+ and thus it implicitly sets the minimum
bound for racecar.

This is an alternative to #256.

Fixes #252