WIP: Thread based concurrency
Concurrently runs a number of 'Racecar::Runner' instances in a fixed size thread pool.

Each thread starts a single `Racecar::Runner` with `Racecar::Consumer`
class instance. All threads run the same consumer class, have the same config
and consume partitions from the same topic(s).

`ThreadPoolRunnerProxy` can be combined with `ParallelRunner`, to run
forks and threads. `ParallelRunner` is not used or battle tested (at
Zendesk) and so this is still not a recommended thing to do.

Other inclusions:

- Signal handling has been moved up one level to the CLI
- Runner-like object interface standardized to `#run` `#stop` `#running?`
- Test can be run locally with Docker, `export LOCAL=1`
- Some test bugs have been fixed, connections now always close and
  orphaned processes raise an exception