Resume topic/partition in all consumers
Prior to this, when a pause expires in a partition that is no longer
assigned to the client, we don't call resume on the rdkafka consumers.
If this partition is then re-assigned to the consumer, we don't start
consuming messages from it because it is still paused as far as the
rdkafka consumer is concerned.

Since we only track `Pause`s by topic and partition, and we don't have
any information about which consumer the pause exists for, our only
option is to resume the topic/partition on all consumers. This won't
cause errors because it's safe to call resume for a non-existent
topic/partition. If there are multiple consumers with the same
topic/partition (like from different Kafka clusters), this has the
potential to erroneously resume partitions. In general this isn't a huge
deal because if the partition fails again it will just be re-paused.
Ultimately it would probably be a good idea to make pauses aware of the
consumer, or add logic to remove rdkafka-level consumer pauses in
rebalance callbacks.